refactor: move collected artists privacy screen to screens folder
